refactor: moves boardEditor module into its own folder
