+++ src/__tests__/pieces/pawn.moves.test.tsx
@@ -3,7 +3,7 @@ import { Coordinate } from '../../utils/board/board.types';
 import {
   createBoard,
   createFromConcisePosition,
-} from '../../utils/board/boardEditor';
+} from '../../utils/board/editor/boardEditor';
 import { PieceColor } from '../../utils/pieces.types';
 import {
   assertCandidateMoves,
