+++ src/__tests__/pieces/support/king.moves.data.ts
@@ -1,4 +1,4 @@
-import { createBoard } from '../../../utils/board/boardEditor';
+import { createBoard } from '../../../utils/board/editor/boardEditor';
 import { BoardAndMoves } from '../common.test.utils';
 
 export const pureKingPositionsAndMoves: BoardAndMoves[] = [
